stactools_pipelines/pipelines/global_mangrove_watch/requirements.txt
stactools-global-mangrove-watch==0.1.0
aws-lambda-powertools
boto3
requests

